norme ok, build ready
