+++ src/execution/process_command.c
@@ -6,7 +6,7 @@
 /*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/10 13:07:11 by jteissie          #+#    #+#             */
-/*   Updated: 2024/07/15 13:34:53 by jteissie         ###   ########.fr       */
+/*   Updated: 2024/07/15 15:05:10 by jteissie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,17 +68,15 @@ int	redirect_parent(int p_fd[])
 	return (dup_status);
 }
 
-int	process_command(t_lex_parser *p, char **envp, t_parser *d, int std_fds[])
+int	process_command(t_lex_parser *p, char **env, t_parser *d, int std_fd[])
 {
 	int			pipe_fd[2];
 	int			has_pipe[2];
-	int			dup_status;
 	t_cmd_table	*cmd_table;
 	pid_t		pid_child;
 
 	cmd_table = p->table;
 	has_pipe[0] = FALSE;
-	dup_status = 0;
 	has_pipe[1] = FALSE;
 	if (open_pipes(p, pipe_fd, has_pipe) == -1)
 		return (PANIC);
@@ -87,10 +85,9 @@ int	process_command(t_lex_parser *p, char **envp, t_parser *d, int std_fds[])
 		return (PANIC);
 	if (pid_child == 0)
 	{
-
-		if (redirect_child(p, pipe_fd, has_pipe, std_fds) == PANIC)
+		if (redir_child(p, pipe_fd, has_pipe, std_fd) == PANIC)
 			handle_error("syscall error in exec child.\n", errno);
-		execute_cmd(cmd_table->cmd, envp, d);
+		execute_cmd(cmd_table->cmd, env, d);
 	}
 	else
 	{
