+++ src/get_prompt.c
@@ -1,0 +1,56 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   get_prompt.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/07/15 14:41:50 by jteissie          #+#    #+#             */
+/*   Updated: 2024/07/15 14:42:10 by jteissie         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+char	*get_trunc_path(void)
+{
+	char	*full_cwd;
+	char	*trunc_cwd;
+	size_t	i;
+	size_t	len;
+	int		depth;
+
+	full_cwd = getcwd(NULL, 0);
+	len = ft_strlen(full_cwd);
+	i = len;
+	depth = 0;
+	while (--i > 0 && depth < 2)
+		if (full_cwd[i] == '/')
+			depth++;
+	if (i == 0)
+		return (full_cwd);
+	trunc_cwd = ft_substr(full_cwd, i + 1, len - i - 1);
+	free(full_cwd);
+	return (trunc_cwd);
+}
+
+char	*get_prompt(char *orig_prompt)
+{
+	char	*trunc_cwd;
+	char	*pre_prompt;
+	char	*prompt;
+
+	if (orig_prompt)
+		free(orig_prompt);
+	pre_prompt = NULL;
+	trunc_cwd = get_trunc_path();
+	if (!trunc_cwd)
+		return (ft_strdup("minishell> "));
+	pre_prompt = ft_strjoin(trunc_cwd, " | minishell> ");
+	free(trunc_cwd);
+	if (!pre_prompt)
+		return (NULL);
+	prompt = ft_strjoin("~", pre_prompt);
+	free(pre_prompt);
+	return (prompt);
+}
